Moved common events to ConnectionController
